refactor(app): move bocx selector options
