+++ app/react/kubernetes/applications/CreateView/placementTypes.tsx
