+++ app/react/kubernetes/configs/CreateView/options.tsx
