+++ app/react/portainer/init/InitAdminView/restore-options.tsx
